Update lib.py
Added logic for graceful copy of TB_sceneOpened.js from avalon-core to TOONBOOM_GLOBAL_SCRIPT_LOCATION if the env already exists and the file is different. The previous logic would clobber an existing env, and therefore any custom scripts a studio had in that location would not be available on launch. 
I suggest putting very clearly the fact that 'TB_sceneOpened.js' will be overwritten in the TOONBOOM_GLOBAL_SCRIPT_LOCATION in documentation, and advising that TDs put their studio startup logic in either their own fork of avalon-core, adding to the TB_sceneOpened.js, in a Harmony Package, or use the TB_sceneOpenedUI.js instead.